refactor: the music track leaves App as a hook
The audio cluster had the cleanest seam left in App: eleven names, and the
whole group reads only four things it does not own - `media`, `dispatchMedia`,
`cuts` and `setLinkPrompt`. That is because of what audio does not touch. It
knows nothing about cuts, layers, the canvas or the timeline; where it sits and
what part of it plays already live in `mediaReducer`. What was still in App was
the browser side.

`useAudioTrack` takes it: the <audio> element, the AudioContext the export
recorder taps, and the two extra shapes of the same sound - a base64 dataURL so
an .emv carries the music, and a Blob cached against that dataURL because
IndexedDB can hold a Blob and an autosave debounce cannot afford base64.

`detachMedia` goes to `core/mediaEl.js` on the way, since App and the hook both
need it and neither owns it - the same reason `core/api.js` exists.

One dependency array gains `audioRef`. The linter could see it was a ref while
it was declared in App and cannot now; a ref object's identity never changes,
so listing it costs nothing and keeps the warning count from growing.

App.jsx: 3945 -> 3868 lines.
